+++ src/redturtle/prenotazioni/browser/prenotazioni_context_state.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 import itertools
 import json
-import logging
 from datetime import date
 from datetime import datetime
 from datetime import timedelta
@@ -16,6 +15,7 @@ from six.moves import range
 
 from redturtle.prenotazioni import _
 from redturtle.prenotazioni import get_or_create_obj
+from redturtle.prenotazioni import logger
 from redturtle.prenotazioni import tznow
 from redturtle.prenotazioni.adapters.booker import IBooker
 from redturtle.prenotazioni.adapters.conflict import IConflictManager
@@ -28,8 +28,6 @@ from redturtle.prenotazioni.content.prenotazione_type import PrenotazioneType
 from redturtle.prenotazioni.utilities.dateutils import hm2DT
 from redturtle.prenotazioni.utilities.dateutils import hm2seconds
 from redturtle.prenotazioni.utilities.urls import urlify
-
-logger = logging.getLogger(__name__)
 
 
 class PrenotazioniContextState(BrowserView):
@@ -316,8 +314,10 @@ class PrenotazioniContextState(BrowserView):
         slot_min_size: seconds
         """
         # we have some conditions to check
+        # XXX: ma questo non si poteva fare prima... molto prima... di arrivare fin qui ?
         if not self.is_valid_day(day):
             return []
+        # XXX: ma questo non si poteva fare prima... molto prima... di arrivare fin qui ?
         if self.maximum_bookable_date:
             if day > self.maximum_bookable_date.date():
                 return []