+++ src/redturtle/prenotazioni/utilities/__init__.py
@@ -1,0 +1,2 @@
+from .email import send_email  # noqa
+from .log_errors import handle_exception_by_log  # noqa